use only http post to save data and merged near-identical modules into one.
