+++ localserver.js
@@ -25,10 +25,7 @@ var routers = [views,
                 auth,
                 fileSystem,
                 require('./server_modules/sessions'),
-                require('./server_modules/teams'),
-                require('./server_modules/settings'),
-                require('./server_modules/stages'),
-                require('./server_modules/scores'),
+                require('./server_modules/data'),
                 require('./server_modules/challenges')];
