--- server_modules/teams.js
@@ -1,29 +1,0 @@
-var utils = require('./utils');
-var fileSystem = require('./file_system');
-
-exports.route = function(app) {
-
-    //get the teams info
-    app.get('/teams',function(req,res) {
-        fileSystem.readJsonFile(fileSystem.getDataFilePath('teams.json')).then(function(result) {
-            res.json(result);
-        }).catch(err => utils.sendError(res, err)).done();
-    });
-
-    app.get('/teams/:nr',function(req,res) {
-        fileSystem.readJsonFile(fileSystem.getDataFilePath('teams.json')).then(function(result) {
-            var team = result.filter(function(team) {
-                return team.number == req.params.nr;
-            })[0];
-            res.json(team);
-        }).catch(err => utils.sendError(res, err)).done();
-    });
-    app.post('/teams/save', function (req, res, next) {
-        var teams = JSON.stringify(req.body.teams);
-        fileSystem.writeFile(fileSystem.getDataFilePath('teams.json'), teams).then(function () {
-            res.status(200).end();
-        }).catch(function (err) {
-            utils.sendError(res, { status: 500, message: `error writing file` })
-        });
-    });
-};
